integration tests

+++ consensus/ibft/src/main/java/tech/pegasys/pantheon/consensus/ibft/payload/IbftMessage.java
@@ -17,5 +17,5 @@ public interface IbftMessage {
 
   Signature getSignature();
 
-  long getMessageType();
+  int getMessageType();
 }
